messed with stuff, bullets now have a constant speed

+++ source/Emitter.cpp
@@ -41,10 +41,10 @@ void Emitter::load()
 			//Create rates to update particles
 			tempParticle->maxLifeTime = rand() % (int)particleFlyweight->maxLifeTime + particleFlyweight->minLifeTime;
 			tempParticle->currLifeTime = 0;
-			tempParticle->colorRateA = (((float)particleFlyweight->startColor.alpha - (float)particleFlyweight->endColor.alpha) / tempParticle->maxLifeTime);
-			tempParticle->colorRateR = (((float)particleFlyweight->startColor.red - (float)particleFlyweight->endColor.red) / tempParticle->maxLifeTime);
-			tempParticle->colorRateG = (((float)particleFlyweight->startColor.green - (float)particleFlyweight->endColor.green) / tempParticle->maxLifeTime);
-			tempParticle->colorRateB = (((float)particleFlyweight->startColor.blue - (float)particleFlyweight->endColor.blue) / tempParticle->maxLifeTime);
+			tempParticle->colorRateA = (((int)particleFlyweight->startColor.alpha - (int)particleFlyweight->endColor.alpha) / tempParticle->maxLifeTime);
+			tempParticle->colorRateR = (((int)particleFlyweight->startColor.red -	(int)particleFlyweight->endColor.red) / tempParticle->maxLifeTime);
+			tempParticle->colorRateG = (((int)particleFlyweight->startColor.green - (int)particleFlyweight->endColor.green) / tempParticle->maxLifeTime);
+			tempParticle->colorRateB = (((int)particleFlyweight->startColor.blue -	(int)particleFlyweight->endColor.blue) / tempParticle->maxLifeTime);
 			tempParticle->scale = particleFlyweight->startScale;
 			tempParticle->scaleRateX = ((particleFlyweight->startScale.width - particleFlyweight->endScale.width) / tempParticle->maxLifeTime);
 			tempParticle->scaleRateY = ((particleFlyweight->startScale.height - particleFlyweight->endScale.height) / tempParticle->maxLifeTime);
@@ -76,13 +76,13 @@ void Emitter::load()
 			{
 				//MAFF very powerful
 				//NOTE:width == radius height is not used
-				float radius = rand() % (int)size.width / 2;
+				float radius = (float)(rand() % (int)size.width / 2);
 				float x = size.width / 2;
 				float y = size.width / 2;
 				//Point for the center of the emitter
 				SGD::Point center = SGD::Point(x, y);
 				//randomize the angle for the circle
-				float randAngle = rand() % 360;
+				float randAngle = (float)(rand() % 360);
 				randAngle = (randAngle / 180)*SGD::PI;
 				//put in the angle
 				tempParticle->position.x = center.x + (radius*cosf(randAngle));
@@ -156,11 +156,11 @@ void Emitter::Update(float dt)
 				case 1://circle
 				{
 					//NOTE:width == radius height is not used
-					float radius = rand() % (int)size.width / 2;
+					float radius = (float)(rand() % (int)size.width / 2);
 					//Point for the center of the emitter
 					SGD::Point center = SGD::Point(position.x + size.width / 2, position.y + size.width / 2);
 					//randomize the angle for the circle
-					float randAngle = rand() % 360;
+					float randAngle = (float)(rand() % 360);
 					//convert to radians
 					randAngle = (randAngle / 180)*SGD::PI;
 					//put in the angle
